Make Base Nav Component for reusability and Nav for Unregistered User
