+++ components/Nav/index.js
@@ -1,3 +1,4 @@
 import Nav from './Nav'
+import NavUser from './NavUser'
 
-export {Nav}
+export {Nav, NavUser}
