+++ pages/index.js
@@ -1,7 +1,7 @@
 import Head from 'next/head'
 import {Container} from '@chakra-ui/react'
 
-import {Nav} from '@components/Nav'
+import {NavUser} from '@components/Nav'
 
 export default function Home() {
   return (
@@ -13,7 +13,7 @@ export default function Home() {
       </Head>
 
       <header>
-        <Nav />
+        <NavUser />
       </header>
     </Container>
   )
